Implementation changes
