+++ common/src/main/kotlin/flank/common/Json.kt
@@ -1,0 +1,5 @@
+package flank.common
+
+import org.json.JSONObject
+
+fun Map<*, *>.toJSONObject() = JSONObject(this)
