+++ tool/analytics/src/main/kotlin/flank/tool/analytics/IgnoreInStatistics.kt
@@ -1,0 +1,3 @@
+package flank.tool.analytics
+
+annotation class IgnoreInStatistics
